SSCS-10818 Moving HearingType to Hearings
