+++ src/main/java/uk/gov/hmcts/reform/sscs/model/single/hearing/HearingDetails.java
@@ -7,8 +7,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import uk.gov.hmcts.reform.sscs.ccd.domain.AmendReason;
 import uk.gov.hmcts.reform.sscs.model.HearingLocation;
+import uk.gov.hmcts.reform.sscs.model.hmc.reference.HearingType;
 import uk.gov.hmcts.reform.sscs.reference.data.model.HearingChannel;
-import uk.gov.hmcts.reform.sscs.reference.data.model.HearingTypeLov;
 
 import java.util.List;
 
@@ -22,7 +22,7 @@ public class HearingDetails {
 
     private boolean autolistFlag;
 
-    private HearingTypeLov hearingType;
+    private HearingType hearingType;
 
     private HearingWindow hearingWindow;
