+++ src/test/resources/serviceHearingValues.json
@@ -2,7 +2,7 @@
   "caseName": "Jane Smith vs DWP",
   "caseNamePublic": null,
   "autoListFlag": false,
-  "hearingType": "SUBSTANTIVE",
+  "hearingType": "BBA3-SUB",
   "caseType": "Benefit",
   "caseCategories": null,
   "duration": 45,
